Overhaul groups.
Instead of one giant group config file, break it down to multiple packages each with a group.
It seemed most of the groups were just to model rooms in my house, with the devices in them.
So, I put each of them in their own package. To make it easy to figure out rooms vs other packages,
I made packages/rooms/ which totally works.

For rooms, I tried to do:

- lights
- switches
- media players
- temperature & presence detection

I also moved cloud configuration to an alexa package. I figured there were some groups that were just conveniences to be able to ask alexa to turn on/off something.